pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "claude-memory"
version = "0.1.0"
description = "Persistent memory system for Claude Code with multi-signal retrieval"
requires-python = ">=3.12"
license = "MIT"
dependencies = [
    "mcp>=1.0.0",
    "sentence-transformers>=3.0.0",
    "sqlite-vec>=0.1.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-ulid>=3.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[project.scripts]
claude-memory = "claude_memory.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/claude_memory"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
